ztpan/requirements.txt
aiogram==3.0.0
aiomysql==0.2.0
asyncpg==0.28.0
python-dotenv==1.0.0
